--- a/src/ps3eye_capi.h
+++ b/src/ps3eye_capi.h
@@ -1,5 +1,5 @@
 /**
- * PS3EYEDriver C API Interface for use with PS Move API
+ * PS3EYEDriver C API Interface
  * Copyright (c) 2014 Thomas Perl <m@thp.io>
  * All rights reserved.
  *