I've been working (with help and testing by @cboulay and @brendanwalker) on improving the performance of the PS3EYEDriver. While performance is probably sufficient on high-spec desktop, there were quite a few reports of different people (including me) having varying performance (sometimes good, sometimes bad) on different machines.

Investigation revealed the cause was two-fold:

1. The driver doesn't submit (nearly) enough USB bulk requests to saturate the bus; the bus is idle the majority of the time. Since the USB protocol is host-initiated (that is, the device will not send any data unless the host asks for it), this also means that the camera will not transfer data to the host for large amounts of time (the host does not ask it for data quick enough to maintain a stable data flow).

2. The code running the libusb event loop needs to be tight (and continous) enough to ensure that transfers are completed and rescheduled in a timely manner. This is easy to overlook; really the only way to get performance right if you're doing this yourself is to have a seperate thread just running the libusb event loop continously.

To fix these issues I've:

- Increased the number of concurrent USB transfers active from 2 to 8. I've tested this on three different machines (low, medium, high spec), and this seems to hit the sweet spot for raw USB performance (max throughput)

- Introduced a thread internal to PS3EYEDriver that is responsible for running the libusb event loop. This frees clients from having to think about it themselves; getting frames from the camera to the host will always be as fast as possible, but getting frames from the driver to the client app is up to the client calling the frame polling function frequently/fast enough.

In a test app I wrote that just does the following, the performance improves dramatically:

``` C++
while (num_frames_captured < num_frames_to_capture)
{
	BeginFrame("Main");
	psmove_tracker_update_image(tracker);
	EndFrame();
	++num_frames_captured;
}
```

Some side-by-side comparisons:

## USB throughput
**Before optimization**
![before_usb](https://cloud.githubusercontent.com/assets/8319699/12530840/cdc909f2-c1eb-11e5-8b44-f0c76cfffec7.jpg)

**After optimization**
![after_usb](https://cloud.githubusercontent.com/assets/8319699/12530859/7e68bd84-c1ec-11e5-957b-d01b603ef3e4.jpg)

As you can see, before optimization, USB throughput barely gets above 20 MB/s, but is very erratic with a lot of drops to even lower (16 MB/s). After optimization, it's stable at 36 MB/s. 

Note that both captures are from the same program capturing 1000 frames. In the second screenshot it finishes in about 16 seconds, which is what you would expect at 60 FPS (1000 frames * 16ms = ~16 seconds). In the first screenshot, the program takes way longer to finish (the end is out of the picture, but I believe it was around the 44 second mark).

## Frame rate (WPA) analysis
**Before optimization**
![before_wpa](https://cloud.githubusercontent.com/assets/8319699/12530841/d1d64ac8-c1eb-11e5-976a-9dc9ab04b941.png)

**After optimization**
![after_wpa](https://cloud.githubusercontent.com/assets/8319699/12530880/24cf97ce-c1ed-11e5-9b93-bac5ab6d4cd7.png)

(Note that the displayed time range and scale are identical across both captures)

The diamonds at the top represent frame markers (they are the BeginFrame() calls you see in my sample code). The middle chart represents CPU usage; a solid color means the thread is currently scheduled in and running, white means the process was scheduled out (waiting for an event).

As you can see, about 30 FPs is the best we get before optimization, but there are a lot of (very large) frame drops, leading to a very erratic framerate. The thread is frequently scheduled out, because it's waiting for data to arrive.

After optimization, the picture is much better: a frame marker every ~16.66 ms for a very solid 60 FPS. The main thread is continously busy, except for some very small white ticks; that's actually where the main thread runs faster than the camera thread, so it has to wait for data to become available.

## Notes
- API Changes
   - `PS3EYECam::handleEvents` has been removed. It is now called by the internal updater thread.
   - `PS3EYECam::getLastFramePointer` has been renamed to `PS3EYECam::getFrame`. In addition, the behaviour has changed: 
     - It now returns a `memcpy`'d buffer that the user is responsible for `free`ing after he's done with it (ie. this is an ownership transfer)
     - Returns immediately if a frame is available, but blocks if not
   - `PS3EYECam::isNewFrame()` has been removed, since `PS3EYECam::getFrame` will now always return a new frame.
- I've renamed ps3eyedriver.h/.cpp to ps3eye_capi.h/.cpp to make it clearer what the difference is with ps3eye.h/.cpp (and to make upstream acceptance more likely).
- These changes make PS3EYEDriver incompatible with your version of PSMoveAPI. I've updated PSMoveAPI seperately (along with some other useful changes) and will open a PR for that separately. **UPDATE** here: https://github.com/thp/psmoveapi/pull/200
- I'm also making this same PR upstream here: https://github.com/inspirit/PS3EYEDriver/pull/21
- There is a pretty long comment thread [here](https://github.com/cboulay/psmove-ue4/issues/18#issuecomment-177260206) that outlines different approaches tried during optimization, which may be interesting to you.

Please let me know if you have questions or want something done differently.

Thanks.