@thp 2016-02-01
Merging here, but has also been merged upstream, so we might as well get rid of this repo at some point.